feat: 2026-09-17 업데이트를 반영한다 — 소울 상시 공격력·소울 잠재·아르고 호의 가호
09-17 표본 472명이 정수 일치 13명으로 무너졌다. 셋을 넣으면 427명 + 유니온 결손 20명.
- 소울 게이지 효과가 상시 적용으로 바뀌어 soul_pad/soul_mad 로 온다 → 무기 고정 공격력·마력.
- 소울 증폭으로 붙는 소울 잠재 soul_potential_option_1~3 → 잠재처럼 줄 단위로. 화면의
  아이템 창에 "소울 잠재능력" 블럭으로 보인다.
- "공격력 +4.5%" 같은 소수 % 가 생겨 % 시트를 double 로 바꾼다 (정수로 자르면 0.2% 어긋남).
- 이벤트 스킬 "아르고 호의 가호"(보공 20%) 를 반영 목록에 넣는다.
history v9 · datasheet v15.

+++ src/main/java/org/whitedoggy/mapleweb2/domain/common/stat/StatSheet.java
@@ -36,15 +36,16 @@ public class StatSheet {
     private int ATTACK_POWER = 0;
     private int MAGIC_POWER = 0;
 
-    private int STR_PERCENT = 0;
-    private int DEX_PERCENT = 0;
-    private int INT_PERCENT = 0;
-    private int LUK_PERCENT = 0;
-    private int HP_PERCENT = 0;
-    private int ALL_STAT_PERCENT = 0;
+    // % 는 소수가 온다 — 소울 잠재 "공격력 +4.5%", "LUK +0.5%" (2026-09-17). 정수로 두면 잘린다.
+    private double STR_PERCENT = 0.0;
+    private double DEX_PERCENT = 0.0;
+    private double INT_PERCENT = 0.0;
+    private double LUK_PERCENT = 0.0;
+    private double HP_PERCENT = 0.0;
+    private double ALL_STAT_PERCENT = 0.0;
 
-    private int ATTACK_POWER_PERCENT = 0;
-    private int MAGIC_POWER_PERCENT = 0;
+    private double ATTACK_POWER_PERCENT = 0.0;
+    private double MAGIC_POWER_PERCENT = 0.0;
 
     private double DAMAGE = 0.0;
     private double BOSS_DAMAGE = 0.0;
